sdk: Move edit timeline unit tests to a separate module
